+++ src/main/java/dinkplugin/notifiers/LootNotifier.java
@@ -39,6 +39,7 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.EnumSet;
 import java.util.List;
+import java.util.Objects;
 import java.util.OptionalDouble;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.CopyOnWriteArraySet;
@@ -80,6 +81,7 @@ public class LootNotifier extends BaseNotifier {
         itemNameAllowlist.addAll(
             ConfigUtil.readDelimited(config.lootItemAllowlist())
                 .map(Utils::regexify)
+                .filter(Objects::nonNull)
                 .collect(Collectors.toList())
         );
 
@@ -87,6 +89,7 @@ public class LootNotifier extends BaseNotifier {
         itemNameDenylist.addAll(
             ConfigUtil.readDelimited(config.lootItemDenylist())
                 .map(Utils::regexify)
+                .filter(Objects::nonNull)
                 .collect(Collectors.toList())
         );
 
@@ -122,6 +125,7 @@ public class LootNotifier extends BaseNotifier {
         itemNames.addAll(
             ConfigUtil.readDelimited(value)
                 .map(Utils::regexify)
+                .filter(Objects::nonNull)
                 .collect(Collectors.toList())
         );
     }