+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Minor: Add `%COUNT%` template variable for loot notifier to display killcount when rich embeds are disabled. (#840)
 - Minor: Allow custom author badge image in Discord rich embeds. (#834)
 - Minor: Add configurable minimum level to warrant a screenshot for skill notifier. (#833)
 - Minor: Add drop source denylist for loot notifier. (#835)